`@remotion/studio-protocol`: Separate protocol capabilities
